+++ src/main/java/com/tiyb/tev/controller/TEVStagingController.java
@@ -15,6 +15,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -24,12 +25,14 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.tiyb.tev.datamodel.staging.StagingPost;
+import com.tiyb.tev.exception.InvalidConvoParentException;
 import com.tiyb.tev.exception.ResourceNotFoundException;
 import com.tiyb.tev.repository.staging.StagingPostRepository;
 
 /**
  * This is the REST controller for working with items to be staged for eventual
- * export.
+ * export. All APIs have the blog set, even in cases where it wouldn't
+ * technically be needed.
  * 
  * @author tiyb
  *
@@ -71,13 +74,15 @@ public class TEVStagingController {
 	private String ERROR_COPYING_FILE;
 
 	/**
-	 * Returns all "staged posts" that have been added to the staging area
+	 * Returns all "staged posts" that have been added to the staging area for a
+	 * given blog
 	 * 
+	 * @param blog Blog for which to return posts
 	 * @return {@link java.util.List List} of all "staged post" IDs in the repo
 	 */
-	@GetMapping("/posts")
-	public List<Long> getAllPosts() {
-		List<StagingPost> posts = stagingRepo.findAll();
+	@GetMapping("/posts/{blog}")
+	public List<Long> getAllPostsForBlog(@PathVariable("blog") String blog) {
+		List<StagingPost> posts = stagingRepo.findByBlog(blog);
 
 		List<Long> listOfIDs = new ArrayList<Long>();
 
@@ -89,46 +94,55 @@ public class TEVStagingController {
 	}
 
 	/**
-	 * Adds a "staged post" to the staging area
+	 * Adds a "staged post" to the staging area for a given blog
 	 * 
+	 * @param blog   The blog for which the staged post should be added
 	 * @param postID The ID of the StagingPost to be saved into the staging area
 	 * @return The same {@link com.tiyb.tev.datamodel.staging.StagingPost
 	 *         StagingPost} object
 	 */
-	@PostMapping("/posts/{id}")
-	public StagingPost createStagedPost(@PathVariable(value = "id") Long postID) {
+	@PostMapping("/posts/{blog}/{id}")
+	public StagingPost createStagedPostForBlog(@PathVariable("blog") String blog, @PathVariable("id") Long postID) {
 		StagingPost post = new StagingPost();
 		post.setId(postID);
+		post.setBlog(blog);
 
 		return stagingRepo.save(post);
 	}
 
 	/**
-	 * Removes a "staged post" from the staging area
+	 * Removes a "staged post" from the staging area for a given blog
 	 * 
-	 * @param id The ID of the "staged post" to be removed
+	 * @param blog Used for validation
+	 * @param id   The ID of the "staged post" to be removed
 	 * @return {@link org.springframework.http.ResponseEntity ResponseEntity} with
 	 *         the details
 	 */
-	@DeleteMapping("/posts/{id}")
-	public ResponseEntity<?> deleteStagedPost(@PathVariable(value = "id") Long id) {
+	@DeleteMapping("/posts/{blog}/{id}")
+	public ResponseEntity<?> deleteStagedPostForBlog(@PathVariable("blog") String blog, @PathVariable("id") Long id) {
 		StagingPost post = stagingRepo.findById(id)
 				.orElseThrow(() -> new ResourceNotFoundException("StagedPost", "id", id));
 
+		if (!blog.equals(post.getBlog())) {
+			logger.error("Invalid staged post/blog combination; post=" + id + ", blog=" + blog);
+			throw new InvalidConvoParentException();
+		}
 		stagingRepo.delete(post);
 
 		return ResponseEntity.ok().build();
 	}
 
 	/**
-	 * Removes all "staged posts" from the staging area
+	 * Removes all "staged posts" from the staging area for a given blog
 	 * 
+	 * @param blog Blog for which to remove the staged posts
 	 * @return {@link org.springframework.http.ResponseEntity ResponseEntity} with
 	 *         the details
 	 */
-	@DeleteMapping("/posts")
-	public ResponseEntity<?> deleteAllStagedPosts() {
-		stagingRepo.deleteAll();
+	@Transactional
+	@DeleteMapping("/posts/{blog}")
+	public ResponseEntity<?> deleteAllStagedPostsForBlog(@PathVariable("blog") String blog) {
+		stagingRepo.deleteByBlog(blog);
 
 		return ResponseEntity.ok().build();
 	}
@@ -153,20 +167,21 @@ public class TEVStagingController {
 	 * files; no error is thrown (though a warning is logged).
 	 * </p>
 	 * 
+	 * @param blog               Blog to which the post exists
 	 * @param postID             The ID of the post for which images should be
 	 *                           exported
 	 * @param pathForDestination The path to which the images should be copied
 	 * @return {@link org.springframework.http.ResponseEntity ResponseEntity} with
 	 *         the details
 	 */
-	@PostMapping("/posts/{id}/exportImages")
-	public ResponseEntity<?> exportImagesForPost(@PathVariable(value = "id") Long postID,
-			@RequestBody String pathForDestination) {
-		if (!postController.fixPhotosForPost(postID)) {
+	@PostMapping("/posts/{blog}/{id}/exportImages")
+	public ResponseEntity<?> exportImagesForBlogForPost(@PathVariable("blog") String blog,
+			@PathVariable("id") Long postID, @RequestBody String pathForDestination) {
+		if (!postController.fixPhotosForBlogForPost(blog, postID)) {
 			return new ResponseEntity<String>("Error getting images for post", null, HttpStatus.FAILED_DEPENDENCY);
 		}
 
-		String imageDirectory = mdController.getMetadata().getBaseMediaPath();
+		String imageDirectory = mdController.getMetadataForBlogOrDefault(blog).getBaseMediaPath();
 		if (imageDirectory.charAt(imageDirectory.length() - 1) != '/') {
 			imageDirectory = imageDirectory + "/";
 		}
